pyproject.toml
[project]
name = "induction-motor-fault-diagnosis"
version = "0.1.0"
description = "ML model for diagnosing induction motor faults from manually entered data"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=1.26.0",
    "pandas>=2.3.0",
    "scikit-learn>=1.7.0",
    "matplotlib>=3.10.3",
    "jupyter>=1.0.0",
    "seaborn>=0.13.0",
    "fpdf>=1.7.2",
    "openpyxl>=3.1.2",
    "lime>=0.2.0.1",
    "flask-login>=0.6.3",
    "tk",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "flake8>=7.3.0",
    "pytest>=8.4.2",
]
